requirements.txt
Django==4.1.7
django-bootstrap-v5==1.0.11
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
psycopg2-binary==2.9.5
